Switch to custom Go application
This is version 2.

We're replacing nginx with a small custom Go app which, for now, just
mimics the previous behaviour, but leaves us with significantly more
flexibility for options going forward.

For one, this now allows, through `SECURE=1`, forcing redirects to
HTTPS, which is a nice little thing and can prevent double redirects if
you redirect to HTTP only to be redirected again to HTTPS.

This also brings down the container size considerably, as we can
statically-build the binary and don't need a full web server and OS:
7.6MB from the previous 130MB (17x smaller!).