Switch to a config file
This is a pretty enormous change which touches every corner of this
project. To make this more generally useful, we want to allow more
flexible redirects to any host, not just a subdomain.

To facilitate this, we need a more robust configuration paradigm than
flat environment variables. A YAML file like a natural fit.

By using a YAML-based config file, we can support rich configuration, as
well as leave the door open for further per-redirect configuration in
the future if we ever need more advanced features.

As part of this, we also want to rewrite the README to focus less on how
to deploy this and more on how to use the actual program itself. The
bits on Docker and Kubernetes were a bit off-topic for a README. We can
resurrect the Kubernetes examples later as a proper in-repo example if
we really want it.